+++ packages/api/lambdas/sf-event-sqs-to-db-records/index.js
@@ -5,27 +5,14 @@ const get = require('lodash/get');
 const AggregateError = require('aggregate-error');
 
 const { parseSQSMessageBody, sendSQSMessage } = require('@cumulus/aws-client/SQS');
-const { describeExecution } = require('@cumulus/aws-client/StepFunctions');
+
 const log = require('@cumulus/common/log');
 const {
   getKnexClient,
-  tableNames,
 } = require('@cumulus/db');
 const {
   getMessageExecutionArn,
-  getExecutionUrlFromArn,
 } = require('@cumulus/message/Executions');
-const {
-  getMessageGranules,
-  getGranuleStatus,
-} = require('@cumulus/message/Granules');
-const {
-  getMessagePdrName,
-  messageHasPdr,
-} = require('@cumulus/message/PDRs');
-const {
-  getMetaStatus,
-} = require('@cumulus/message/workflows');
 const Execution = require('../../models/executions');
 const Granule = require('../../models/granules');
 const Pdr = require('../../models/pdrs');
@@ -45,147 +32,9 @@ const {
   writePdr,
 } = require('./write-pdr');
 
-const writeGranuleViaTransaction = async ({
-  cumulusMessage,
-  granule,
-  collectionCumulusId,
-  providerCumulusId,
-  executionCumulusId,
-  pdrCumulusId,
-  trx,
-}) =>
-  trx(tableNames.granules)
-    .insert({
-      granule_id: granule.granuleId,
-      status: getGranuleStatus(cumulusMessage, granule),
-      collection_cumulus_id: collectionCumulusId,
-      provider_cumulus_id: providerCumulusId,
-      execution_cumulus_id: executionCumulusId,
-      pdr_cumulus_id: pdrCumulusId,
-    });
-
-/**
- * Write a granule to DynamoDB and Postgres
- *
- * @param {Object} params
- * @param {Object} params.granule - An API granule object
- * @param {Object} params.cumulusMessage - A workflow message
- * @param {string} params.collectionCumulusId
- *   Cumulus ID for collection referenced in workflow message, if any
- * @param {string} params.executionCumulusId
- *   Cumulus ID for execution referenced in workflow message, if any
- * @param {Knex} params.knex - Client to interact with Postgres database
- * @param {string} [params.providerCumulusId]
- *   Cumulus ID for provider referenced in workflow message, if any
- * @param {string} [params.pdrCumulusId]
- *   Cumulus ID for PDR referenced in workflow message, if any
- * @param {Object} [params.granuleModel]
- *   Optional override for the granule model writing to DynamoDB
- *
- * @returns {Promise}
- * @throws
- */
-const writeGranule = async ({
-  granule,
-  cumulusMessage,
-  executionDescription,
-  executionUrl,
-  collectionCumulusId,
-  executionCumulusId,
-  knex,
-  providerCumulusId,
-  pdrCumulusId,
-  granuleModel,
-}) =>
-  knex.transaction(async (trx) => {
-    await writeGranuleViaTransaction({
-      cumulusMessage,
-      granule,
-      collectionCumulusId,
-      providerCumulusId,
-      executionCumulusId,
-      pdrCumulusId,
-      trx,
-    });
-    return granuleModel.storeGranuleFromCumulusMessage({
-      granule,
-      cumulusMessage,
-      executionUrl,
-      executionDescription,
-    });
-  });
-
-/**
- * Write granules to DynamoDB and Postgres
- *
- * @param {Object} params
- * @param {Object} params.cumulusMessage - A workflow message
- * @param {string} params.collectionCumulusId
- *   Cumulus ID for collection referenced in workflow message, if any
- * @param {string} params.executionCumulusId
- *   Cumulus ID for execution referenced in workflow message, if any
- * @param {Knex} params.knex - Client to interact with Postgres database
- * @param {string} [params.providerCumulusId]
- *   Cumulus ID for provider referenced in workflow message, if any
- * @param {string} [params.pdrCumulusId]
- *   Cumulus ID for PDR referenced in workflow message, if any
- * @param {Object} [params.granuleModel]
- *   Optional override for the granule model writing to DynamoDB
- *
- * @returns {Promise<Object[]>}
- *  true if there are no granules on the message, otherwise
- *  results from Promise.allSettled for all granules
- * @throws {Error} - if no collection is provided
- */
-const writeGranules = async ({
-  cumulusMessage,
-  collectionCumulusId,
-  executionCumulusId,
-  knex,
-  providerCumulusId,
-  pdrCumulusId,
-  granuleModel = new Granule(),
-}) => {
-  if (!collectionCumulusId) {
-    throw new Error('Collection reference is required for granules');
-  }
-
-  const granules = getMessageGranules(cumulusMessage);
-  const executionArn = getMessageExecutionArn(cumulusMessage);
-  const executionUrl = getExecutionUrlFromArn(executionArn);
-
-  let executionDescription;
-  try {
-    executionDescription = await describeExecution({ executionArn });
-  } catch (error) {
-    log.error(`Could not describe execution ${executionArn}`, error);
-  }
-
-  // Process each granule in a separate transaction via Promise.allSettled
-  // so that they can succeed/fail independently
-  const results = await Promise.allSettled(granules.map(
-    (granule) => writeGranule({
-      granule,
-      cumulusMessage,
-      executionDescription,
-      executionUrl,
-      collectionCumulusId,
-      providerCumulusId,
-      executionCumulusId,
-      pdrCumulusId,
-      knex,
-      granuleModel,
-    })
-  ));
-  const failures = results.filter((result) => result.status === 'rejected');
-  if (failures.length > 0) {
-    const allFailures = failures.map((failure) => failure.reason);
-    const aggregateError = new AggregateError(allFailures);
-    log.error('Failed writing some granules to Dynamo', aggregateError);
-    throw aggregateError;
-  }
-  return results;
-};
+const {
+  writeGranules,
+} = require('./write-granules');
 
 const writeRecordsToDynamoDb = async (cumulusMessage) => {
   const executionModel = new Execution();
@@ -277,7 +126,5 @@ const handler = async (event) => {
 
 module.exports = {
   handler,
-  writeGranuleViaTransaction,
-  writeGranules,
   writeRecords,
 };
