+++ packages/api/tests/lambdas/sf-event-sqs-to-db-records/test-index.js
@@ -29,8 +29,6 @@ const shouldWriteExecutionStub = sandbox.stub().resolves(true);
 
 const {
   handler,
-  writeGranules,
-  writePdr,
   writeRecords,
 } = proxyquire('../../../lambdas/sf-event-sqs-to-db-records', {
   '@cumulus/aws-client/SQS': {
@@ -221,162 +219,6 @@ test.after.always(async (t) => {
   await t.context.knex.destroy();
   await t.context.knexAdmin.raw(`drop database if exists "${t.context.testDbName}"`);
   await t.context.knexAdmin.destroy();
-});
-
-test('writeGranules() throws an error if collection is not provided', async (t) => {
-  const { cumulusMessage, knex, providerCumulusId } = t.context;
-  await t.throwsAsync(
-    writeGranules({
-      cumulusMessage,
-      collectionCumulusId: undefined,
-      providerCumulusId,
-      knex,
-    })
-  );
-});
-
-test('writeGranules() saves granule records to Dynamo and RDS if RDS write is enabled', async (t) => {
-  const {
-    cumulusMessage,
-    granuleModel,
-    knex,
-    collectionCumulusId,
-    providerCumulusId,
-    granuleId,
-  } = t.context;
-
-  await writeGranules({
-    cumulusMessage,
-    collectionCumulusId,
-    providerCumulusId,
-    knex,
-  });
-
-  t.true(await granuleModel.exists({ granuleId }));
-  t.true(
-    await doesRecordExist({ granule_id: granuleId }, knex, tableNames.granules)
-  );
-});
-
-test('writeGranules() handles successful and failing writes independently', async (t) => {
-  const {
-    cumulusMessage,
-    granuleModel,
-    knex,
-    collectionCumulusId,
-    providerCumulusId,
-    granuleId,
-  } = t.context;
-
-  const granule2 = {
-    // no granule ID should cause failure
-  };
-  cumulusMessage.payload.granules = [
-    ...cumulusMessage.payload.granules,
-    granule2,
-  ];
-
-  await t.throwsAsync(writeGranules({
-    cumulusMessage,
-    collectionCumulusId,
-    providerCumulusId,
-    knex,
-  }));
-
-  t.true(await granuleModel.exists({ granuleId }));
-  t.true(
-    await doesRecordExist({ granule_id: granuleId }, knex, tableNames.granules)
-  );
-});
-
-test('writeGranules() throws error if any granule writes fail', async (t) => {
-  const {
-    cumulusMessage,
-    knex,
-    collectionCumulusId,
-    providerCumulusId,
-  } = t.context;
-
-  cumulusMessage.payload.granules = [
-    ...cumulusMessage.payload.granules,
-    // this object is not a valid granule, so its write should fail
-    {},
-  ];
-
-  await t.throwsAsync(writeGranules({
-    cumulusMessage,
-    collectionCumulusId,
-    providerCumulusId,
-    knex,
-  }));
-});
-
-test.serial('writeGranules() does not persist records to Dynamo or RDS if Dynamo write fails', async (t) => {
-  const {
-    cumulusMessage,
-    granuleModel,
-    knex,
-    collectionCumulusId,
-    providerCumulusId,
-    granuleId,
-  } = t.context;
-
-  const fakeGranuleModel = {
-    storeGranuleFromCumulusMessage: () => {
-      throw new Error('Granules dynamo error');
-    },
-  };
-
-  const [error] = await t.throwsAsync(
-    writeGranules({
-      cumulusMessage,
-      collectionCumulusId,
-      providerCumulusId,
-      knex,
-      granuleModel: fakeGranuleModel,
-    })
-  );
-
-  t.true(error.message.includes('Granules dynamo error'));
-  t.false(await granuleModel.exists({ granuleId }));
-  t.false(
-    await doesRecordExist({ granule_id: granuleId }, knex, tableNames.granules)
-  );
-});
-
-test.serial('writeGranules() does not persist records to Dynamo or RDS if RDS write fails', async (t) => {
-  const {
-    cumulusMessage,
-    granuleModel,
-    knex,
-    collectionCumulusId,
-    providerCumulusId,
-    granuleId,
-  } = t.context;
-
-  const fakeTrxCallback = (cb) => {
-    const fakeTrx = sinon.stub().returns({
-      insert: () => {
-        throw new Error('Granules RDS error');
-      },
-    });
-    return cb(fakeTrx);
-  };
-  const trxStub = sinon.stub(knex, 'transaction').callsFake(fakeTrxCallback);
-  t.teardown(() => trxStub.restore());
-
-  const [error] = await t.throwsAsync(writeGranules({
-    cumulusMessage,
-    collectionCumulusId,
-    providerCumulusId,
-    knex,
-  }));
-
-  t.true(error.message.includes('Granules RDS error'));
-  t.false(await granuleModel.exists({ granuleId }));
-  t.false(
-    await doesRecordExist({ granule_id: granuleId }, knex, tableNames.granules)
-  );
 });
 
 test.serial('writeRecords() only writes records to Dynamo if shouldWriteExecutionToRDS returns false', async (t) => {
